requirements.txt
itsdangerous
discord.py==2.7.1
requests==2.33.1
python-dotenv==1.2.2
asyncpraw==7.8.1
aiohttp==3.13.5
asyncpg==0.31.0
pytz==2026.1
parsedatetime==2.6
openai==2.31.0
pydantic==2.12.5
motor==3.7.1
apscheduler==3.11.2
PyYAML>=6.0
loguru==0.7.3
pydantic_settings>=2.13.0